Update emailharvester.py
For security reasons, and to counteract MITM attacks in this project and implementation of Advanced SSL Bypassing, SSL certificate pinning techniques have been applied. Here, SSL certificate pinning makes client to continue interacting with trusted server only by comparing the server’s public key fingerprint with that of a previously stored one. This method brings the SSL connections authenticity by ensuring that the server has not been interfered. Conveyor makes use of ‘urllib3’ for showing HTTPS connection and SSL, the validation used in `certifi` which is a list of Mozilla’s trusted root certificate. Also for getting the certificate fingerprint server public key is extracted and then hashed using the `cryptography` library. In this context, the fingerprint used is compared to the pinned version and; if they do not match, the connection is terminated to discourage unauthorized entry. In order to eliminate the SSL verification warnings, there is a urllib3 setting, which can be disabled as `urllib3.disable_warnings()` but this should be avoided for usage in tests on live systems. This approach facilitates secure and trusted media when participating in tests and practices flexibility.